requirements.txt
click==8.1.7
SQLAlchemy==2.0.36
python-dotenv==1.0.1
setuptools==75.2.0
psycopg2==2.9.10
flake8==7.1.1
spacy==3.8.2
